backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pypdf==3.17.1
python-docx==1.1.0
chromadb==0.4.18
sentence-transformers>=2.2.2
scikit-learn==1.3.2
rapidfuzz==3.5.2
openai>=1.12.0
pandas==2.1.3
python-dotenv==1.0.0
werkzeug==3.0.1
reportlab==4.0.6
